refactor: Rename file (DSD-151)
